requirements.txt
flask
numpy
torch
torchvision
Image
opencv-python
pillow
FlickrAPI
model_selection